Intel RealSense 2 Update
Removing some useless code and now records to a folder using automatic file recording names.

Some bugs I have noticed -- Recording using the current methodology keeps everything in memory. This is not a good idea for programs that need to record for a long time. I need to find an alternative solution for this...